Check if options are good before execution.

+++ ics2qtcal.sh
@@ -12,15 +12,20 @@ shift `expr $OPTIND - 1`
 icsfile="$1"
 caldb="$2"
 
-echo "Creating temporary copy of $icsfile with valid lines"
-# Create a copy and remove X-MOZ-LASTACK lines that are not understood by Tie::iCal
-grep -v X-MOZ-LASTACK "$icsfile" >"${icsfile}.tmp"
-# Normalize line endings to unix format (line feed only)
-sed -e 's/\r$//g' "${icsfile}.tmp" >"${icsfile}-2.tmp"
+if [ -n $icsfile ] && [ -n $caldb ] then
+    echo "Creating temporary copy of $icsfile with valid lines into db $caldb"
+    # Create a copy and remove X-MOZ-LASTACK lines that are not understood by Tie::iCal
+    grep -v X-MOZ-LASTACK "${icsfile}" >"${icsfile}.tmp"
+    # Normalize line endings to unix format (line feed only)
+    sed -e 's/\r$//g' "${icsfile}.tmp" >"${icsfile}-2.tmp"
 
-echo "Processing file $icsfile"
-perl ics2qtcal.pl "$verbose" --ical "${icsfile}-2.tmp" --qtopiadb "$caldb" --notesdirectory "./Annotator-tmp"
+    echo "Processing file $icsfile"
+    perl ics2qtcal.pl "$verbose" --ical "${icsfile}-2.tmp" --qtopiadb "$caldb" --notesdirectory "./Annotator-tmp"
 
-echo "Deleting temporary files"
-rm "${icsfile}.tmp"
-rm "${icsfile}-2.tmp"
+    echo "Deleting temporary files"
+    rm "${icsfile}.tmp"
+    rm "${icsfile}-2.tmp"
+else
+    echo "ERROR : unable to treat empty files ! Command was (without options) : \n"
+    echo "$0 $*"
+fi
